+++ CHANGELOG.rst
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Current
+-------
+
+- Nothin yet
+
+
 0.4.2
 -----