+++ CHANGELOG.rst
@@ -1,8 +1,8 @@
 Changelog
 =========
 
-Current
--------
+0.5.0
+-----
 
 - Allow shorter syntax to set operation id: ``@api.doc('my-operation')``
 - Added a shortcut to specify the expected input model: ``@api.expect(my_fields)``
@@ -10,6 +10,7 @@ Current
 - Added ``@api.extend()`` to extend models
 - Ensure coherence between ``required`` and ``allow_null`` for ``NestedField``
 - Support list of primitive types and list of models as body
+- Upgraded to latest version of Swagger UI
 - Fixes